slider light refactoring
